move stories to stories/Label
