+++ packages/react-components/react-label/src/stories/Label/LabelDefault.stories.tsx
@@ -1,4 +1,4 @@
 import * as React from 'react';
-import { Label, LabelProps } from '../index';
+import { Label, LabelProps } from '../../index';
 
 export const Default = (props: LabelProps) => <Label {...props}>This is a label</Label>;
