+++ packages/react-components/react-label/src/stories/Label/LabelWeight.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Label } from '../index';
+import { Label } from '../../index';
 
 export const Weight = () => <Label weight="semibold">Strong label</Label>;
